Refactored User Repository and wrote initial test cases
In this commit, I refactored the User Repository by creating it and setting up a TODO list of possible test cases. The interface UserRepository is not currently being used since it was just created, and the private field entityManager is assigned but not accessed, which is okay since no tests have been initialized yet. Additionally, I wrote initial test cases for the repository using unit testing techniques.